requirements.txt
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==10.4.0
